refactor: refactor of ProfileManager with pattern Repository/Adapter and Single Sesponsibility principle
