+++ src/main/scala/users_management/repository/MongoProfileManager.scala
@@ -1,0 +1,2 @@
+package users_management.repository
+
